refactor: remove actions crate; move main crate to not be in workspace
