+++ backend/api/Makefile.toml
@@ -1,0 +1,58 @@
+###################
+## Config        ##
+###################
+
+[config]
+skip_core_tasks = true
+default_to_workspace = false
+
+###################
+## Test          ##
+###################
+
+[tasks.test]
+command = "cargo"
+args = ["test"]
+
+[tasks.test-watch]
+command = "cargo"
+args = ["watch", "-w", "../../shared/rust", "-w", ".", "-x", "test --features local-quiet -- --nocapture"]
+
+###################
+## Dev           ##
+###################
+
+[tasks.dev-proxy]
+run_task = { name = [ "google-sql-proxy", "server-local-quiet-proxy", ], parallel = true }
+
+
+[tasks.dev-warnings-proxy]
+run_task = { name = [ "google-sql-proxy", "server-local-proxy", ], parallel = true }
+
+
+[tasks.server-local-proxy]
+command = "systemfd"
+args = ["--no-pid", "-s", "http::8081", "--", "cargo", "watch", "-w", "../../config/rust/src", "-w", "../../shared/rust/src", "-w", "../_core/rust/src", "-w", ".", "-x", "run --features local,sqlproxy"]
+
+[tasks.server-local-quiet-proxy]
+command = "systemfd"
+args = ["--no-pid", "-s", "http::8081", "--", "cargo", "watch", "-w", "../../config/rust/src", "-w", "../../shared/rust/src", "-w", "../_core/rust/src", "-w", ".", "-x", "run --features local-quiet,sqlproxy"]
+
+[tasks.dev]
+run_task = { name = [ "server-local-quiet"] }
+
+[tasks.dev-warnings]
+run_task = { name = [ "server-local"] }
+
+
+[tasks.server-local]
+command = "systemfd"
+args = ["--no-pid", "-s", "http::8081", "--", "cargo", "watch", "-w", "../../config/rust/src", "-w", "../../shared/rust/src", "-w", "../_core/rust/src", "-w", ".", "-x", "run --features local"]
+
+[tasks.server-local-quiet]
+command = "systemfd"
+args = ["--no-pid", "-s", "http::8081", "--", "cargo", "watch", "-w", "../../config/rust/src", "-w", "../../shared/rust/src", "-w", "../_core/rust/src", "-w", ".", "-x", "run --features local-quiet"]
+
+[tasks.google-sql-proxy]
+env = { CARGO_MAKE_SCRIPT_FORCE_PIPE_STDIN = true }
+script = ["cd ../../../build-utils","npm run cloud-sql-proxy"]
