--- backend/api/crates/actions/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "actions"
-version = "0.1.0"
-authors = ["dakom <david.komer@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-once_cell = "1.4.0"
-percent-encoding = "2.1.0"
-simplelog = "^0.8"
-jsonwebtoken = "^7.1"
-reqwest = { version = "^0.10", features = ["json"] }
-cfg-if = "^0.1"
-sqlx = { version = "0.4.0-beta.1", default-features = false, features = [ "runtime-tokio", "macros", "postgres", "uuid", "json", "chrono", "offline"] }
-chrono = { version = "^0.4", features = ["serde"] }
-serde = { version = "^1.0", features = ["derive"] }
-serde_json = "^1.0"
-serde_derive = "^1.0"
-strum = "^0.18"
-strum_macros = "^0.18"
-lazy_static = "^1.4"
-log = "^0.4"
-futures-util = "0.3.4"
-rand = "0.7"
-shared = {path = "../../../../shared/rust"}
-core = {path = "../../../_core/rust"}
-config = {path = "../../../../config/rust"}
-openssl = { version = "^0.10", features = ["vendored"], optional = true }
-
-[features]
-local = ["core/local"]
-release = ["core/release", "openssl"]
-sandbox = ["core/sandbox", "openssl"]
-quiet = []
