--- backend/api/main/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-name = "ji-cloud-api"
-version = "0.1.0"
-authors = ["dakom <david.komer@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-tokio = { version = "^0.2", features = ["full"] }
-simplelog = "^0.8"
-cfg-if = "^0.1"
-log = "^0.4"
-dotenv = "^0.15"
-once_cell = "1.4.0"
-percent-encoding = "2.1.0"
-listenfd = {version = "^0.3", optional = true }
-jsonwebtoken = "^7.1"
-reqwest = { version = "^0.10", features = ["json"] }
-sqlx = { version = "0.4.0-beta.1", default-features = false, features = [ "runtime-tokio", "macros", "postgres", "uuid", "json", "chrono", "offline"] }
-chrono = { version = "^0.4", features = ["serde"] }
-serde = { version = "^1.0", features = ["derive"] }
-serde_json = "^1.0"
-serde_derive = "^1.0"
-strum = "^0.18"
-strum_macros = "^0.18"
-lazy_static = "^1.4"
-futures-util = "0.3.4"
-rand = "0.7"
-shared = {path = "../../../shared/rust"}
-config = {path = "../../../config/rust"}
-core = {path = "../../_core/rust"}
-actions = {path = "../crates/actions"}
-openssl = { version = "^0.10", features = ["vendored"], optional = true }
-actix-cors = "0.3.0-alpha.1"
-actix-rt = "1.0.0"
-actix-service = "1.0.5"
-actix-web = "3.0.0-beta.1"
-anyhow = "1.0.32"
-futures = "0.3.5"
-
-[features]
-local = ["listenfd", "core/local", "actions/local"]
-release = ["core/release", "actions/release", "openssl"]
-sandbox = ["core/sandbox", "actions/sandbox", "openssl"]
-quiet = ["actions/quiet"]
-sqlproxy = []
-local-quiet = ["local", "quiet"]
